+++ R/featureLayers.R
@@ -119,6 +119,29 @@ featureLayerOptions <- function(
 #' @param fitBounds Whether to set the maps bounds to fit the data in the featureLayer
 #'   options
 #' @export
+#' @examples
+#' leaflet() %>%
+#'   addEsriBasemapLayer(esriBasemapLayers$Streets) %>%
+#'   setView(-122.667, 45.526, 13) %>%
+#'   addEsriFeatureLayer(
+#'     url = paste0("https://services.arcgis.com/rOo16HdIMeOBI4Mb/arcgis/rest/services/",
+#'                  "Heritage_Trees_Portland/FeatureServer/0"),
+#'     useServiceSymbology = TRUE,
+#'     labelProperty = "COMMON_NAM", labelOptions = labelOptions(textsize = "12px"),
+#'   popupProperty = JS(paste0(
+#'     "function(feature) {",
+#'     "  return L.Util.template(",
+#'     "    \"<h3>{COMMON_NAM}</h3><hr />",
+#'     "      <p>This tree is located at {ADDRESS} and its scientific name is {SCIENTIFIC}.</p>",
+#'     "    \",",
+#'     "    feature.properties",
+#'     "  );",
+#'     "}"
+#'   )))
+#'
+#' ## for more examples see
+#' # browseURL(system.file("examples/featureLayers.R", package = "leaflet.esri"))
+#' # browseURL(system.file("examples/multipleFeatureLayers.R", package = "leaflet.esri"))
 addEsriFeatureLayer <- function(
   map, url, useServiceSymbology = FALSE,
   options = featureLayerOptions(),